requirements.txt
# Document generation
feedgen
pyyaml

# File storage
boto3

# YouTube
youtube-dl
